+++ src/css/getCSSValue/index.ts
@@ -7,7 +7,7 @@ export type TGetCSSValueReturn = ReturnType<typeof getCSSValue>;
 /**
  * Gets a calculated CSS property of an DOM-element
  * @param el{HTMLElement} DOM element
- * @param prop{String} CSS property
+ * @param prop{string} CSS property
  * @see https://developer.mozilla.org/en-US/docs/Web/API/Window/getComputedStyle
  * @returns {string}
  * @throws {TypeError} getCSSValue: el must be an HTMLElement