+++ src/css/isSelectorValid/index.ts
@@ -6,7 +6,7 @@ export type TIsSelectorValidReturn = ReturnType<typeof isSelectorValid>;
 
 /**
  * Checks if string is valid CSS selector
- * @param str{String} source selector
+ * @param str{string} source selector
  * @returns {boolean}
  * @throws {TypeError} isSelectorValid: str must be a string
  * @see https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_selectors