+++ src/css/setCSSVar/index.ts
@@ -7,8 +7,8 @@ export type TSetCSSVarReturn = ReturnType<typeof setCSSVar>;
 /**
  * Sets CSS3 variable to specific DOM node
  * @param el{HTMLElement=} DOM element
- * @param variable{String} variable name
- * @param value{String|Number|Boolean=} variable value
+ * @param variable{string} variable name
+ * @param value{string|number|boolean=} variable value
  * @see https://developer.mozilla.org/en-US/docs/Web/CSS/Using_CSS_custom_properties
  * @returns {void}
  * @throws {TypeError} setCSSVar: el must be an HTMLElement